feat: migrate platform_id to BigInteger for NHN int64 IDs

+++ db/models/application.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Optional
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy import Integer, String, DateTime, ForeignKey, UniqueConstraint
+from sqlalchemy import BigInteger, Integer, String, DateTime, ForeignKey, UniqueConstraint
 from db.models.base import Base
 from services.wanted.wanted_constants import WANTED
 
@@ -10,7 +10,7 @@ class Application(Base):
     __tablename__ = "applications"
     internal_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     source: Mapped[str] = mapped_column(String(20), nullable=False, default=WANTED)
-    platform_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    platform_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     job_id: Mapped[int] = mapped_column(ForeignKey("jobs.internal_id"), nullable=False)
     status: Mapped[str] = mapped_column(String(50), nullable=False)
     apply_time: Mapped[Optional[datetime]] = mapped_column(DateTime)
